+++ src/services/base-filters-parser.js
@@ -11,7 +11,7 @@ const parseFiltersString = (filtersString) => {
 };
 
 // NOTICE: Apply the formatCondition function to a condition (leaf).
-const parseCondition = (condition, formatCondition) => {
+const parseCondition = async (condition, formatCondition) => {
   if (_.isEmpty(condition)) { throw new InvalidFiltersFormat('Empty condition in filter'); }
   if (!_.isObject(condition)) { throw new InvalidFiltersFormat('Condition cannot be a raw value'); }
   if (_.isArray(condition)) { throw new InvalidFiltersFormat('Filters cannot be a raw array'); }
@@ -25,28 +25,30 @@ const parseCondition = (condition, formatCondition) => {
 
 // NOTICE: Call the formatAggregation function on the node and propagate it to its childs or
 //         call the parseCondition function on the node if the node is a leaf.
-const parseAggregation = (node, formatAggregation, formatCondition) => {
+const parseAggregation = async (node, formatAggregation, formatCondition) => {
   if (_.isEmpty(node)) { throw new InvalidFiltersFormat('Empty condition in filter'); }
   if (!_.isObject(node)) { throw new InvalidFiltersFormat('Filters cannot be a raw value'); }
   if (_.isArray(node)) { throw new InvalidFiltersFormat('Filters cannot be a raw array'); }
 
   if (!node.aggregator) { return parseCondition(node, formatCondition); }
 
-  const formatedConditions = [];
 
   if (!_.isArray(node.conditions)) {
     throw new InvalidFiltersFormat('Filters\' conditions must be an array');
   }
 
-  node.conditions.forEach(condition => formatedConditions
-    .push(parseAggregation(condition, formatAggregation, formatCondition)));
+  const promises = [];
+  node.conditions.forEach(condition =>
+    promises.push(parseAggregation(condition, formatAggregation, formatCondition)));
+
+  const formatedConditions = await Promise.all(promises);
 
   return formatAggregation(node.aggregator, formatedConditions);
 };
 
 // NOTICE: Recursively call the formatAggregation function on the nodes of the filters tree and
 //         propagate the formatCondition function to the leaves.
-const perform = (filtersString, formatAggregation, formatCondition) => {
+const perform = async (filtersString, formatAggregation, formatCondition) => {
   const filters = parseFiltersString(filtersString);
 
   if (!filters) { return null; }
@@ -63,8 +65,8 @@ const aggregateAssociations = (_aggregator, conditionsAssociations) =>
   _.flatten(conditionsAssociations);
 
 // NOTICE: Recursively populate the associations names from the filters.
-const getAssociations = (filtersString) => {
-  const associations = perform(
+const getAssociations = async (filtersString) => {
+  const associations = await perform(
     filtersString,
     aggregateAssociations,
     getConditionAssociation,